langgraph_service/requirements.txt
# LangGraph Service Dependencies
# Main application dependencies
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
pydantic==2.5.0

# LangGraph and AI dependencies
langgraph
groq

# Testing dependencies
pytest==7.4.3
requests==2.31.0
